Set up initial backend structure with FastAPI (#496)
* Initial setup for sqlite without flask-sqlalchemy dependency

* Add wrapper for fastapi's APIRouter to handle trailing slashes

* Move static files serving to FastAPI

* Move commits and projects packages views from Flask to FastAPI

* Change main app from Flask to FastAPI

* Move tags, dashboards and dashboard_apps views from Flask to FastAPI

* Add CORS middleware for FastAPI app

* Change sqlalchemy session usage to more fastapi way

* Change db upgrade and gunicorn(changed to hypercorn) commands construction for FastAPI

* Remove "manage.py" file to get rid of flask-migrate, flask-script dependencies

* Restructure web dir and cleanup requirements

* Change FastAPI server from hypercorn to uvicorn and make the corresponding updates on project requirements
