+++ aim/cli/up/utils.py
@@ -1,8 +1,7 @@
-import sys
 import os
 import click
 
-from aim.engine.configs import AIM_FLASK_ENV_KEY
+from aim.engine.configs import AIM_WEB_ENV_KEY
 
 
 def repo_init_alert():
@@ -13,19 +12,20 @@ def repo_init_alert():
 def build_db_upgrade_command():
     from aim import web
     web_dir = os.path.dirname(web.__file__)
-    manage_file = os.path.join(web_dir, 'manage.py')
     migrations_dir = os.path.join(web_dir, 'migrations')
-    return [sys.executable, manage_file,
-            'db', 'upgrade', '--directory', migrations_dir]
+    if os.getenv(AIM_WEB_ENV_KEY) == 'dev':
+        ini_file = os.path.join(migrations_dir, 'alembic_dev.ini')
+    else:
+        ini_file = os.path.join(migrations_dir, 'alembic.ini')
+    return ['alembic', '-c', ini_file, 'upgrade', 'head']
 
 
-def build_gunicorn_command(host, port, num_workers):
-    bind_address = "%s:%s" % (host, port)
-    cmd = ['gunicorn', '-b', bind_address, '-w', '%s' % num_workers,
-           '--timeout', '300', '--graceful-timeout', '300']
-    if os.getenv(AIM_FLASK_ENV_KEY) == 'dev':
-        cmd += ['--reload', '--log-level', 'info']
+def build_uvicorn_command(host, port, num_workers):
+    cmd = ['uvicorn', '--host', host, '--port', '%s' % port, '--workers', '%s' % num_workers]
+    if os.getenv(AIM_WEB_ENV_KEY) == 'dev':
+        import aim
+        cmd += ['--reload', '--reload-dir', os.path.dirname(aim.__file__), '--log-level', 'debug']
     else:
         cmd += ['--log-level', 'error']
-    cmd += ['aim.web.run']
+    cmd += ['aim.web.run:app']
     return cmd
