+++ aim/engine/configs.py
@@ -48,7 +48,7 @@ DEFAULT_SYSTEM_TRACKING_INT = 10
 AIM_UI_DEFAULT_PORT = 43800
 AIM_UI_DEFAULT_HOST = '127.0.0.1'
 AIM_TF_LOGS_PATH_KEY = '__AIM_TF_LOGS_PATH__'
-AIM_FLASK_ENV_KEY = '__AIM_FLASK_ENV__'
+AIM_WEB_ENV_KEY = '__AIM_WEB_ENV__'
 AIM_UI_MOUNTED_REPO_PATH = '__AIM_UI_MOUNT_REPO_PATH__'
 AIM_UI_TELEMETRY_KEY = 'AIM_UI_TELEMETRY_ENABLED'
