+++ aim/web/api/dashboard_apps/serializers.py
@@ -1,5 +1,5 @@
-from aim.web.app.dashboard_apps.models import ExploreState
-from aim.web.app.utils import BaseModelSerializer, BaseSerializer, Field, ModelField, json_loads_or_none
+from aim.web.api.dashboard_apps.models import ExploreState
+from aim.web.api.utils import BaseModelSerializer, BaseSerializer, Field, ModelField, json_loads_or_none
 
 
 class ExploreStateModelSerializer(BaseModelSerializer):
