+++ aim/web/api/dashboards/serializers.py
@@ -1,13 +1,14 @@
-from aim.web.app.dashboards.models import Dashboard
-from aim.web.app.dashboard_apps.models import ExploreState
+from aim.web.api.dashboards.models import Dashboard
+from aim.web.api.dashboard_apps.models import ExploreState
 
 
-def dashboard_response_serializer(dashboard_object):
+def dashboard_response_serializer(dashboard_object, session):
     if not isinstance(dashboard_object, Dashboard):
         return None
 
-    app = ExploreState.query.filter(ExploreState.dashboard_id == dashboard_object.uuid,
-                                    ExploreState.is_archived == False).first()
+    app = session.query(ExploreState)\
+        .filter(ExploreState.dashboard_id == dashboard_object.uuid, ExploreState.is_archived == False)\
+        .first()
 
     response = {
         'id': dashboard_object.uuid,
