+++ aim/web/api/utils.py
@@ -1,13 +1,13 @@
 import datetime
 import json
 import math
-import os
 import pytz
-
-from flask import send_from_directory
 
 from collections import OrderedDict
 from collections.abc import Iterable
+from fastapi import APIRouter as FastAPIRouter
+from fastapi.types import DecoratedCallable
+from typing import Any, Callable
 
 
 def datetime_now():
@@ -45,14 +45,29 @@ def unsupported_float_type(value) -> bool:
 
     return False
 
-def send_from_directory_gzip_compressed(directory, filename, **options):
-    compressed_filename = '{}.gz'.format(filename)
-    if os.path.exists(os.path.join(directory, compressed_filename)):
-        rv = send_from_directory(directory, compressed_filename, **options)
-        rv.headers.add('Content-Encoding', 'gzip')
-        return rv
 
-    return send_from_directory(directory, filename, **options)
+class APIRouter(FastAPIRouter):
+    def api_route(
+        self, path: str, *, include_in_schema: bool = True, **kwargs: Any
+    ) -> Callable[[DecoratedCallable], DecoratedCallable]:
+        if path.endswith("/"):
+            path = path[:-1]
+
+        add_path = super().api_route(
+            path, include_in_schema=include_in_schema, **kwargs
+        )
+
+        alternate_path = path + "/"
+        add_alternate_path = super().api_route(
+            alternate_path, include_in_schema=False, **kwargs
+        )
+
+        def decorator(func: DecoratedCallable) -> DecoratedCallable:
+            add_alternate_path(func)
+            return add_path(func)
+
+        return decorator
+
 
 def json_loads_or_none(obj):
     return json.loads(obj) if obj else None
