+++ requirements.txt
@@ -1,23 +1,19 @@
-click>=6.7
+# sdk
+aimrecords==0.0.7
 anytree>=2.8.0
-pyrser>=0.2.0
+click>=7.0
 GitPython>=3.0.4
-requests>=2.0.0
-psutil>=5.6.7
-py3nvml>=0.2.5
-docker>=4.1.0
-aimrecords==0.0.6
 protobuf>=3.11.0
-
-# requirements to add (to be tested yet if something is missing)
-# alembic
-# Flask
-# Flask-Cors
-# Flask-Migrate
-# Flask-RESTful
-# Flask-Script
-# Flask-SQLAlchemy
-# SQLAlchemy
-# pysqlite3
-# pytz
-# tensorboard
+psutil>=5.6.7
+pyrser>=0.2.0
+py3nvml>=0.2.5
+requests>=2.0.0
+# server
+aiofiles>=0.5.0
+alembic>=1.4.0
+fastapi>=0.65.0
+uvicorn>=0.12.0
+python-multipart>=0.0.3
+pytz>=2019.1
+SQLAlchemy>=1.3.0
+tensorboard>=2.0.0
